Rename entity ID format card and move registry settings to a dedicated file
